chore: redis 모듈 추가 #2
refactor: orders 서비스 리펙토링
